add:修复bug

+++ JiaMianWuHui/app/src/main/java/com/qsd/jmwh/module/home/park/bean/HomePersonListBean.java
@@ -43,7 +43,7 @@ public class HomePersonListBean implements Serializable {
         public String sDatePro;
         public int nGalaryCapacity;
         public int nAuthType;
-        public int nOffLineMin;
+        public long nOffLineMin;
         public int nInfoSet;
         public int nOnLine;
         public boolean bHiddenRang;
